Use a shared type file for external types
