+++ draw.ts
@@ -1,31 +1,10 @@
 import { TRIANGULATION } from './triangulation';
-
-export declare type Coords3D = Array<[number, number, number]>;
-export declare type Coords2DPlus = Array<[number, number, ...number[]]>;
-interface FacePrediction {
-    faceInViewConfidence: number;
-    boundingBox: {
-        topLeft: [number, number];
-        bottomRight: [number, number];
-    };
-    mesh: Coords3D;
-    scaledMesh: Coords3D;
-    annotations?: {
-        [key: string]: Coords3D;
-    };
-}
-
-interface HandPrediction {
-    annotations: {
-        [key: string]: Coords3D;
-    };
-    handInViewConfidence: number;
-    landmarks: Coords3D;
-    boundingBox: {
-        topLeft: [number, number];
-        bottomRight: [number, number];
-    };
-}
+import {
+    Coords3D,
+    Coords2DPlus,
+    FacePrediction,
+    HandPrediction
+} from './type';
 
 function drawPoint(ctx: CanvasRenderingContext2D, y: number, x: number, r: number) {
     ctx.beginPath();
